+++ tests/WaveSystem_SphericalExplosion_impl_seq.cxx
@@ -172,7 +172,7 @@ int main(int argc, char *argv[])
 
    // Problem data
     double tmax=0.05;
-    double ntmax=2000000;
+    int ntmax=2000000;
     int freqSortie=1;
     string fileOutPut="SphericalWave";
